+++ src/MetaheuristicsPlatform/Algorithms/Multimodal/SpeciesBasedPso/SpeciesBasedPsoOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.Multimodal.SpeciesBasedPso;
+
+public static class SpeciesBasedPsoOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Li",
+            2004,
+            "Adaptively choosing neighbourhood bests using species in a particle swarm optimizer for multimodal function optimization",
+            "GECCO 2004, LNCS 3102, 105-116",
+            "10.1007/978-3-540-24854-5_10");
+}